requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
psycopg2==2.7.4
PyJWT==1.6.1
pytz==2018.3
requests==2.18.4
urllib3==1.22
